chore: remove obsolete ci/temp_patches
Remove ci/temp_patches, which was obsoleted in 23f608fd once it
was no longer used by the sync script. It should have been
deleted then.

Remove it not only to clean up dead code, but because it contains
a reference to `micro_copts`, which is about to be refactored
away, and we don't want to leave stray references to it in the
tree.