fix(memory_arena_threshold): with TF_LITE_STATIC_MEMORY
Fix the broken build due to redefinition of the threshold when
TF_LITE_STATIC_MEMORY is set. Apparently this case isn't triggered in
any Bazel test, only in Make.

Simplify the threshold specification by only depending on whether
compression is enabled and not also on whether TF_LITE_STATIC_MEMORY is
in use.